Updating to rev681
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/ecohydrology@254 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/SWATmodel/src/SWAT2012/ascrv.f
@@ -44,10 +44,10 @@
 !!    Intrinsic: Log     
 
 !!    ~ ~ ~ ~ ~ ~ END SPECIFICATIONS ~ ~ ~ ~ ~ ~
-      real :: xx
+      real*8 :: xx
 
-      real, intent (in) :: x1, x2, x3, x4
-      real, intent (out) :: x5, x6
+      real*8, intent (in) :: x1, x2, x3, x4
+      real*8, intent (out) :: x5, x6
       
       xx = 0.0
       x5 = 0.0
@@ -56,7 +56,6 @@
       xx = Log(x3/x1 - x3)
       x6 = (xx - Log(x4/x2 - x4)) / (x4 - x3)
       x5 = xx + (x3 * x6)
-
 
       return
       end
