+++ pkg/SWATmodel/src/SWAT2012/orgncswat.f
@@ -63,14 +63,14 @@
 
       integer, intent (in) :: iwave
       integer :: j
-      real :: xx, wt1, er, conc
+      real*8 :: xx, wt1, er, conc
 
       j = 0
       j = ihru
 
       xx = 0.
       wt1 = 0.  !! conversion factor
-      er = 0.      !! enrichment ratio
+      er = 0. !! enrichment ratio
       if (iwave <= 0) then
         !! HRU calculations
         xx = sol_n(1,j) + sol_fon(1,j) + sol_mn(1,j)
@@ -103,9 +103,9 @@
       !! update soil nitrogen pools only for HRU calculations
       if (iwave <= 0 .and. xx > 1.e-6) then
         xx1 = (1. - sedorgn(j) / xx)
-            sol_n(1,j) = sol_n(1,j) * xx1
-            sol_fon(1,j) = sol_fon(1,j) * xx1
-            sol_mn(1,j) = sol_mn(1,j) * xx1
+       sol_n(1,j) = sol_n(1,j) * xx1
+       sol_fon(1,j) = sol_fon(1,j) * xx1
+       sol_mn(1,j) = sol_mn(1,j) * xx1
       end if
 
       return
