+++ pkg/SWATmodel/src/SWAT2012/wndgen.f
@@ -36,13 +36,12 @@
 
       integer, intent (in) :: j
  
-      real :: v6
+      real*8 :: v6
 
       !! Generate wind speed !!
       v6 = 0.
       v6 = Aunif(rndseed(idg(5),j))
       u10(j) = wndav(i_mo,hru_sub(j)) * (-Log(v6)) ** 0.3
       
-
       return
       end
